light-uniform-PTQ/requirements.txt
compressai==1.2.4
numpy==1.19.5
Pillow==8.4.0
Pillow==10.0.1
pytorch_msssim==1.0.0
PyYAML==5.4.1
PyYAML==6.0.1
timm==0.6.13
torch==1.10.0
torchvision==0.11.0
